Import and clean CDS 411 course materials
Notes
-----

The following are the course materials for the Fall 2018 semester of CDS 411
offered at George Mason University, with the README.md file itself serving as
the course webpage. The contents of the repository include the lecture notes in
R Markdown and Jupyter notebook formats, as well as the homework assignments and
final project instructions.

The following minor changes have been made to the course materials to prepare
them for public release.

* References to specific calendar dates, such as due dates for assignments,
  dates on the topic schedule, and dates on the lecture notes, have been
  removed. Some stray dates may still remain here and there, please open an
  issue if you find any that I missed.
* The lecture notes have been renumbered.
* GitHub classroom links for obtaining starter code have been removed.
* Committing files in the build/ directory after running "make all" is no longer
  allowed.