Update topic schedule with links to Jupyter nbviewer renderings
Notes
-----

The links in the topic schedule now point directly to the lecture notes instead
of taking the user to the directory containing the files. Notes contained in a
Jupyter notebook or a PDF file now link to the file via nbviewer.jupyter.org.
Python source files are also now listed on the topic schedule, and clicking the
links will take the user to the file, just within the GitHub interface itself.

The TOC at the top of the README.md file has also been fixed to say "Topic
schedule" instead of "Calendar".